+++ packages/bitte-wallet/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @near-wallet-selector/bitte-wallet
 
+## 9.5.1
+
+## 9.5.0
+
+## 9.4.1
+
 ## 9.4.0
 
 ## 9.3.1