+++ packages/intear-wallet/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @near-wallet-selector/intear-wallet
 
+## 9.5.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @near-wallet-selector/core@9.5.1
+  - @near-wallet-selector/wallet-utils@9.5.1
+
+## 9.5.0
+
+### Minor Changes
+
+- [#1396](https://github.com/near/wallet-selector/pull/1396) [`679b534`](https://github.com/near/wallet-selector/commit/679b5340f82acddf43e5b98c33e07f76605e0d8e) Thanks [@Sliman4](https://github.com/Sliman4)! - Added support for Intear Wallet desktop app or on self-hosted domain
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @near-wallet-selector/core@9.5.0
+  - @near-wallet-selector/wallet-utils@9.5.0
+
+## 9.4.1
+
+### Patch Changes
+
+- Updated dependencies [[`134a872`](https://github.com/near/wallet-selector/commit/134a8723b938cdd922ddbf1eec528cdac7ae6c3e)]:
+  - @near-wallet-selector/core@9.4.1
+  - @near-wallet-selector/wallet-utils@9.4.1
+
 ## 9.4.0
 
 ### Patch Changes